+++ CakeCrafter.Backend/CakeCrafter.API/Controllers/Templates/GenericController.cs
@@ -5,24 +5,29 @@ namespace CakeCrafter.API.Controllers.Templates
 {
     [Route("api/[controller]")]
     [ApiController]
-    public abstract class InformationController<TModel> : ControllerBase where TModel : class
+    public abstract class GenericController<TModel> : ControllerBase where TModel : class
     {
         private readonly IGenericService<TModel> _service;
 
-        public InformationController(IGenericService<TModel> service)
+        public GenericController(IGenericService<TModel> service)
         {
             _service = service;
         }
 
         [HttpGet]
-        public async Task<ActionResult<List<TModel>>> GetModelsAsync()
+        public async Task<ActionResult<List<TModel>>> GetModels()
         {
             return Ok(await _service.Get());
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<TModel>> GetModelByIdAsync(int id)
+        public async Task<ActionResult<TModel>> GetModelById(int id)
         {
+            if (id < 0)
+            {
+                return BadRequest();
+            }
+
             var dbModel = await _service.GetById(id);
             if (dbModel == null)
             {
@@ -32,26 +37,43 @@ namespace CakeCrafter.API.Controllers.Templates
         }
 
         [HttpPost]
-        public async Task<ActionResult<TModel>> AddModelAsync(TModel model)
+        public async Task<ActionResult<TModel>> AddModel(TModel model)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             var result = await _service.Create(model);
             return Ok(result);
         }
 
         [HttpPut("{id}")]
-        public async Task<ActionResult<List<TModel>>> UpdateModelAsync(int id, TModel model)
+        public async Task<ActionResult<List<TModel>>> UpdateModel(int id, TModel model)
         {
+            if (!ModelState.IsValid || id < 1)
+            {
+                return BadRequest();
+            }
+
             var dbModel = await _service.Update(model, id);
+
             if (dbModel == null)
             {
                 return NotFound();
             }
+
             return Ok(dbModel);
         }
 
         [HttpDelete("{id}")]
-        public async Task<ActionResult<List<TModel>>> DeleteModelAsync(int id)
+        public async Task<ActionResult<List<TModel>>> DeleteModel(int id)
         {
+            if (id < 1)
+            {
+                return BadRequest();
+            }
+
             return Ok(await _service.Delete(id));
         }
     }
